Dashboard mostly complete
Styled the dashboard (would appreciate some comments on it). SurveyCard component will be used to fetch and render analytics and other details from database on whatever survey user clicks on in the dashboard.

Added routes in index.js for the dashboard and a survey creator page (I want to try using SurveyJS to create a form builder).

Changed Login.jsx so it redirects to /dashboard instead of /profile (just wanted to note that).